requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.3.7
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
loguru==0.7.2
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0 